Moved authController inside userController
