--- routes/auth.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const authController = require('../controllers/authController');
-const router = express.Router();
-
-router.post('/register', authController.registerUser);
-router.post('/login', authController.login);
-
-module.exports = router;
