+++ routes/index.js
@@ -1,11 +1,8 @@
 const express = require('express');
-const jwtMiddleware = require('../middleware/jwtMiddleware');
-
 const router = express.Router();
-const userRouter = require('./user');
-const authRouter = require('./auth');
 
-router.use('/', authRouter);
-router.use('/user', jwtMiddleware, userRouter);
+const userRouter = require('./user');
+
+router.use('/user', userRouter);
 
 module.exports = router;
